refactor(input): relocate default/tmux input dispatch under src/input/
Move src/mode/default/input.rs → src/input/default_mode.rs and
src/mode/tmux/{input,mouse,mouse/*,forward,gate,pane_hit,window_bar_input}.rs
→ src/input/tmux/ with a new src/input/tmux.rs declaring those submodules.
Update all cross-module use paths; widen TmuxActiveSet, request_detach,
CopyModeSnapshot, cursor_deltas, and cell_at_pane to pub(crate) to support
the new call sites. No logic changes.
