+++ src/input/default_mode.rs
@@ -205,7 +205,7 @@ fn apply_ime_commit_to_terminal(
     });
 }
 
-pub(super) fn should_disable_input(
+pub(crate) fn should_disable_input(
     composing: bool,
     window_focused: bool,
     webview_focused: bool,
