+++ src/input/tmux/mouse.rs
@@ -12,12 +12,12 @@ mod decide;
 mod effect;
 mod webview;
 
-use super::copy_mode::{CopyModeSnapshot, cell_at_pane};
 use super::pane_hit::tmux_pane_at_phys;
-use super::render::{DividerPixelRect, PackedTmuxLayout};
 use crate::configs::OzmuxConfigsResource;
 use crate::input::InputPhase;
 use crate::input::gesture::ClickTracker;
+use crate::mode::tmux::copy_mode::{CopyModeSnapshot, cell_at_pane};
+use crate::mode::tmux::render::{DividerPixelRect, PackedTmuxLayout};
 use crate::ui::copy_mode::CopyModeState;
 use crate::ui::copy_search::CopyPrompt;
 use bevy::ecs::system::SystemParam;
@@ -50,7 +50,7 @@ impl Plugin for MousePlugin {
                     .run_if(pointer_active)
                     .after(tmux_webview_pointer)
                     .in_set(InputPhase::Dispatch)
-                    .in_set(super::TmuxActiveSet),
+                    .in_set(crate::mode::tmux::TmuxActiveSet),
             )
             .add_plugins((webview::WebviewPointerPlugin, apply::ApplyPlugin));
     }
@@ -437,7 +437,7 @@ fn continuation_ctx(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::mode::tmux::pane_hit::tmux_pane_at_phys;
+    use crate::input::tmux::pane_hit::tmux_pane_at_phys;
     use crate::webview_pointer::WebviewPress;
     use bevy::input::ButtonState;
     use bevy::input::mouse::MouseButtonInput;
