+++ src/input/tmux/mouse/effect.rs
