+++ src/main.rs
@@ -25,10 +25,10 @@ use bootstrap::OzmuxBootstrapPlugin;
 use configs::OzmuxConfigsPlugin;
 use font::FontBridgePlugin;
 use input::OzmuxShortcutPlugin;
+use input::default_mode::DefaultHostInputPlugin;
 use input::ime::ImePlugin;
 use input::option_as_alt::OptionAsAltPlugin;
 use mode::default::CopyModeInputPlugin;
-use mode::default::DefaultHostInputPlugin;
 use mode::default::DefaultModePlugin;
 use mode::default::DefaultWebviewPointerPlugin;
 use mode::tmux::OzmuxTmuxPlugin;
